Move scheduler support to standalone package
