+++ src/ansys/fluent/core/scheduler/__init__.py
@@ -1,0 +1,3 @@
+"""A package providing job scheduler support."""
+
+from .load_machines import load_machines  # noqa: F401
